Cleaning up scripts

+++ auto-setup/core/cfssl-certificate-authority/post_install_scripts/installNginx.sh
@@ -28,7 +28,9 @@ server {
 ''' | sudo tee /etc/nginx/sites-available/rabbitmq.conf
 
 # Create shortcut to enable NGINX virtual host
-ln -s /etc/nginx/sites-available/rabbitmq.conf /etc/nginx/sites-enabled/rabbitmq.conf
+if [[ ! -f /etc/nginx/sites-enabled/rabbitmq.conf ]]; then
+  ln -s /etc/nginx/sites-available/rabbitmq.conf /etc/nginx/sites-enabled/rabbitmq.conf
+fi
 
 # Restart NGINX so new virtual host is loaded
 sudo systemctl restart nginx
